Move Select model path
